+++ packages/web-sdk/src/metas/index.ts
@@ -1,5 +1,7 @@
 export { browserMeta } from './browser';
 
+export { osMeta } from './os';
+
 export { createSession } from './session';
 
 export { sdkMeta } from './sdk';